Separate PaymentSelection.IconLoader into it's own type.
