+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/flowcontroller/DefaultFlowControllerTest.kt
@@ -2377,8 +2377,10 @@ internal class DefaultFlowControllerTest {
         activityResultCaller = activityResultCaller,
         statusBarColor = { STATUS_BAR_COLOR },
         paymentOptionFactory = PaymentOptionFactory(
-            resources = context.resources,
-            imageLoader = StripeImageLoader(context),
+            iconLoader = PaymentSelection.IconLoader(
+                resources = context.resources,
+                imageLoader = StripeImageLoader(context),
+            ),
             context = context,
         ),
         paymentOptionCallback = paymentOptionCallback,
